+++ src/gwt/src/org/rstudio/core/client/command/impl/DesktopMenuCallback.java
@@ -92,6 +92,16 @@ public class DesktopMenuCallback implements MenuCallback
       callbacks.setCommandChecked(commandId, checked);
    }-*/;
 
+   public static final void setCommandShortcut(String commandId, String shortcut)
+   {
+      setCommandShortcutImpl(commandId, shortcut, MENU_CALLBACKS);
+   }
+
+   private native static final void setCommandShortcutImpl(String commandId, String shortcut, JavaScriptObject callbacks)
+   /*-{
+      callbacks.setCommandShortcut(commandId, shortcut);
+   }-*/;
+
    public native static final void setMainMenuEnabled(boolean enabled)
    /*-{
        if ($wnd.desktopMenuCallback)